requirements.txt
streamlit
opencv-python
clarifai-grpc